+++ internal/signals.go
@@ -1,0 +1,6 @@
+package internal
+
+import "os"
+
+// TerminationSignals are the signals that cause the program to terminate.
+var TerminationSignals []os.Signal